requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
channels==4.0.0
click==8.1.7
colorama==0.4.6
Django==4.2.11
djangorestframework==3.15.1
environs==11.0.0
frozenlist==1.4.1
gunicorn==21.2.0
h11==0.14.0
idna==3.6
magic-filter==1.0.12
marshmallow==3.21.1
multidict==6.0.5
packaging==24.0
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.1
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
uvicorn==0.29.0
yarl==1.9.4
